refactor: #102 -  회원가입 API 테스트 책임별 분리
