+++ src/app/api/auth/resend-verification-email/route.test.ts
@@ -1,13 +1,22 @@
 import { NextRequest } from "next/server";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
+// API 응답 코드 상수
 import { AUTH_API_CODES } from "@/features/auth/constants/authApiCodes";
+// Rate Limit store 초기화 (테스트 간 상태 격리)
 import { resetResendRateLimitStore } from "@/features/auth/resend-verification-email/lib/checkResendRateLimit";
+// 외부 의존성 (모두 mock 대상)
 import { getLastVerificationResendAt } from "@/features/auth/resend-verification-email/lib/getLastVerificationResendAt";
 import { resendVerificationEmail } from "@/features/auth/resend-verification-email/lib/resendVerificationEmail";
 import { setLastVerificationResendAt } from "@/features/auth/resend-verification-email/lib/setLastVerificationResendAt";
 
+// 테스트 대상 API 핸들러
 import { POST } from "./route";
+
+/**
+ * 외부 의존성 mock
+ * - DB 조회, 외부 이메일 전송, 상태 저장 모두 실제 호출 방지
+ */
 vi.mock(
   "@/features/auth/resend-verification-email/lib/getLastVerificationResendAt",
 );
@@ -18,7 +27,11 @@ vi.mock(
   "@/features/auth/resend-verification-email/lib/setLastVerificationResendAt",
 );
 
-describe("PR-API-09 이메일 인증 재전송 API 성공 흐름", () => {
+describe("이메일 인증 재전송 API 성공 흐름", () => {
+  /**
+   * NextRequest 생성 헬퍼
+   * - 실제 HTTP 요청을 흉내냄
+   */
   function makeRequest(body: object): NextRequest {
     return new NextRequest(
       "http://localhost/api/auth/resend-verification-email",
@@ -31,37 +44,44 @@ describe("PR-API-09 이메일 인증 재전송 API 성공 흐름", () => {
   }
 
   beforeEach(() => {
+    // 모든 mock 초기화
     vi.clearAllMocks();
+
+    // Rate Limit 상태 초기화 (테스트 간 영향 제거)
     resetResendRateLimitStore();
+
+    // 기본 mock 동작 정의
     vi.mocked(getLastVerificationResendAt).mockResolvedValue(null);
     vi.mocked(resendVerificationEmail).mockResolvedValue(undefined);
     vi.mocked(setLastVerificationResendAt).mockResolvedValue(undefined);
   });
 
-  // TC-01: 유효 요청 시 resend 호출 및 성공 응답 반환
+  // TC-01: 정상 흐름
   it("TC-01. 유효한 요청 시 resend를 1회 호출하고 성공 응답을 반환한다", async () => {
     const response = await POST(makeRequest({ email: " Test@Example.COM " }));
     const body = await response.json();
 
-    // Response assertions
+    // 응답 검증 (HTTP + 계약)
     expect(response.status).toBe(200);
     expect(body.success).toBe(true);
     expect(body.code).toBe(AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_SUCCESS);
+
+    // email normalize 확인 (trim + lowercase)
     expect(body.data.email).toBe("test@example.com");
     expect(body.data.resent).toBe(true);
 
-    // External call assertions
+    // 외부 의존 호출 검증
     expect(resendVerificationEmail).toHaveBeenCalledTimes(1);
     expect(resendVerificationEmail).toHaveBeenCalledWith("test@example.com");
 
-    // Security assertions
+    // 보안: 불필요한 필드 노출 금지
     expect(body).not.toHaveProperty("error");
     expect(body).not.toHaveProperty("errors");
   });
 });
 
 describe("PR-API-10 이메일 재전송 1분 쿨다운 검증", () => {
-  const now = 1_700_000_000_000;
+  const now = 1_700_000_000_000; // 고정된 현재 시간 (테스트 안정성)
 
   function makeRequest(): NextRequest {
     return new NextRequest(
@@ -77,18 +97,25 @@ describe("PR-API-10 이메일 재전송 1분 쿨다운 검증", () => {
   beforeEach(() => {
     vi.clearAllMocks();
     resetResendRateLimitStore();
+
+    /**
+     * fake timer 사용 이유:
+     * - 시간 기반 로직(쿨다운)을 deterministic하게 테스트하기 위함
+     */
     vi.useFakeTimers();
     vi.setSystemTime(now);
+
     vi.mocked(getLastVerificationResendAt).mockResolvedValue(null);
     vi.mocked(resendVerificationEmail).mockResolvedValue(undefined);
     vi.mocked(setLastVerificationResendAt).mockResolvedValue(undefined);
   });
 
   afterEach(() => {
+    // 다른 테스트에 영향 주지 않도록 원복
     vi.useRealTimers();
   });
 
-  // TC-01: 마지막 재전송 타임스탬프 없으면 첫 요청 허용
+  // TC-01: 최초 요청 허용
   it("TC-01. 마지막 재전송 타임스탬프가 없으면 첫 요청이 허용된다", async () => {
     vi.mocked(getLastVerificationResendAt).mockResolvedValue(null);
 
@@ -98,21 +125,17 @@ describe("PR-API-10 이메일 재전송 1분 쿨다운 검증", () => {
     expect(response.status).toBe(200);
     expect(body.success).toBe(true);
     expect(body.code).toBe(AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_SUCCESS);
-    expect(typeof body.data).toBe("object");
+
+    // 흐름 검증: 조회 → 전송 → 저장
     expect(getLastVerificationResendAt).toHaveBeenCalledTimes(1);
-    expect(getLastVerificationResendAt).toHaveBeenCalledWith(
-      "test@example.com",
-    );
     expect(resendVerificationEmail).toHaveBeenCalledTimes(1);
-    expect(resendVerificationEmail).toHaveBeenCalledWith("test@example.com");
-    expect(setLastVerificationResendAt).toHaveBeenCalledTimes(1);
     expect(setLastVerificationResendAt).toHaveBeenCalledWith(
       "test@example.com",
       now,
     );
   });
 
-  // TC-02: 59초 전 타임스탬프 → 차단
+  // TC-02: 59초 → 차단
   it("TC-02. 마지막 재전송 타임스탬프가 59초 전이면 요청이 차단된다", async () => {
     vi.mocked(getLastVerificationResendAt).mockResolvedValue(now - 59 * 1000);
 
@@ -124,13 +147,13 @@ describe("PR-API-10 이메일 재전송 1분 쿨다운 검증", () => {
     expect(body.code).toBe(
       AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_COOLDOWN_CONFLICT,
     );
-    expect(body.data).toBeNull();
-    expect(getLastVerificationResendAt).toHaveBeenCalledTimes(1);
+
+    // 차단 시 외부 호출 없음
     expect(resendVerificationEmail).toHaveBeenCalledTimes(0);
     expect(setLastVerificationResendAt).toHaveBeenCalledTimes(0);
   });
 
-  // TC-03: 60초 전 타임스탬프 → 허용
+  // TC-03: 60초 → 허용 (경계값)
   it("TC-03. 마지막 재전송 타임스탬프가 60초 전이면 요청이 허용된다", async () => {
     vi.mocked(getLastVerificationResendAt).mockResolvedValue(now - 60 * 1000);
 
@@ -139,60 +162,42 @@ describe("PR-API-10 이메일 재전송 1분 쿨다운 검증", () => {
 
     expect(response.status).toBe(200);
     expect(body.success).toBe(true);
-    expect(body.code).toBe(AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_SUCCESS);
-    expect(typeof body.data).toBe("object");
-    expect(getLastVerificationResendAt).toHaveBeenCalledTimes(1);
+
     expect(resendVerificationEmail).toHaveBeenCalledTimes(1);
-    expect(resendVerificationEmail).toHaveBeenCalledWith("test@example.com");
-    expect(setLastVerificationResendAt).toHaveBeenCalledTimes(1);
     expect(setLastVerificationResendAt).toHaveBeenCalledWith(
       "test@example.com",
       now,
     );
   });
 
-  // TC-04: 차단 시 외부 resend 호출 없음
+  // TC-04: 차단 시 resend 호출 없음
   it("TC-04. 차단된 요청은 외부 resend 호출을 발생시키지 않는다", async () => {
     vi.mocked(getLastVerificationResendAt).mockResolvedValue(now - 59 * 1000);
 
     const response = await POST(makeRequest());
-    const body = await response.json();
+    await response.json();
 
-    expect(response.status).toBe(409);
-    expect(body.success).toBe(false);
-    expect(body.code).toBe(
-      AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_COOLDOWN_CONFLICT,
-    );
     expect(resendVerificationEmail).toHaveBeenCalledTimes(0);
   });
 
-  // TC-05: 차단 시 타임스탬프 업데이트 없음
+  // TC-05: 차단 시 timestamp 업데이트 없음
   it("TC-05. 차단된 요청은 마지막 재전송 타임스탬프를 업데이트하지 않는다", async () => {
     vi.mocked(getLastVerificationResendAt).mockResolvedValue(now - 59 * 1000);
 
     const response = await POST(makeRequest());
-    const body = await response.json();
+    await response.json();
 
-    expect(response.status).toBe(409);
-    expect(body.success).toBe(false);
-    expect(body.code).toBe(
-      AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_COOLDOWN_CONFLICT,
-    );
     expect(setLastVerificationResendAt).toHaveBeenCalledTimes(0);
   });
 
-  // TC-06: 허용 시 현재 시각으로 타임스탬프 저장
+  // TC-06: 허용 시 timestamp 저장
   it("TC-06. 허용된 요청은 마지막 재전송 타임스탬프를 현재 시각으로 저장한다", async () => {
     vi.mocked(getLastVerificationResendAt).mockResolvedValue(null);
 
     const response = await POST(makeRequest());
-    const body = await response.json();
+    await response.json();
 
-    expect(response.status).toBe(200);
-    expect(body.success).toBe(true);
-    expect(body.code).toBe(AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_SUCCESS);
     expect(resendVerificationEmail).toHaveBeenCalledTimes(1);
-    expect(setLastVerificationResendAt).toHaveBeenCalledTimes(1);
     expect(setLastVerificationResendAt).toHaveBeenCalledWith(
       "test@example.com",
       now,
@@ -200,7 +205,7 @@ describe("PR-API-10 이메일 재전송 1분 쿨다운 검증", () => {
   });
 });
 
-describe("PR-API-11 이메일 재전송 Rate Limit 검증 (5회/1분)", () => {
+describe("이메일 재전송 Rate Limit 검증 (5회/1분)", () => {
   function makeRequest(): NextRequest {
     return new NextRequest(
       "http://localhost/api/auth/resend-verification-email",
@@ -215,8 +220,11 @@ describe("PR-API-11 이메일 재전송 Rate Limit 검증 (5회/1분)", () => {
   beforeEach(() => {
     vi.clearAllMocks();
     resetResendRateLimitStore();
+
+    // Rate limit은 시간 기반 → fake timer 필수
     vi.useFakeTimers();
     vi.setSystemTime(new Date("2026-01-01T00:00:00.000Z"));
+
     vi.mocked(getLastVerificationResendAt).mockResolvedValue(null);
     vi.mocked(resendVerificationEmail).mockResolvedValue(undefined);
     vi.mocked(setLastVerificationResendAt).mockResolvedValue(undefined);
@@ -226,21 +234,19 @@ describe("PR-API-11 이메일 재전송 Rate Limit 검증 (5회/1분)", () => {
     vi.useRealTimers();
   });
 
-  // TC-01: 동일 이메일 5회 이내 모두 허용
+  // TC-01: 5회까지 허용
   it("TC-01. 동일 이메일로 5회 이내 요청은 모두 허용된다", async () => {
     for (let i = 0; i < 5; i++) {
       const response = await POST(makeRequest());
       const body = await response.json();
 
-      expect(response.status).not.toBe(429);
       expect(body.success).toBe(true);
-      expect(body.code).toBe(AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_SUCCESS);
     }
 
     expect(resendVerificationEmail).toHaveBeenCalledTimes(5);
   });
 
-  // TC-02: 6번째 요청은 차단
+  // TC-02: 6번째 차단
   it("TC-02. 동일 이메일로 6번째 요청은 차단된다", async () => {
     for (let i = 0; i < 5; i++) {
       await POST(makeRequest());
@@ -252,12 +258,12 @@ describe("PR-API-11 이메일 재전송 Rate Limit 검증 (5회/1분)", () => {
 
     expect(response.status).toBe(429);
     expect(body.success).toBe(false);
-    expect(body.code).toBe(AUTH_API_CODES.RESEND_RATE_LIMIT_EXCEEDED);
-    expect(body.data).toBeNull();
+
+    // 차단 시 외부 호출 없음
     expect(resendVerificationEmail).toHaveBeenCalledTimes(0);
   });
 
-  // TC-03: 4회 후 5번째는 허용 (경계값)
+  // TC-03: 5번째는 허용 (경계값)
   it("TC-03. 동일 이메일로 4회 요청 후 5번째 요청은 허용된다", async () => {
     for (let i = 0; i < 4; i++) {
       await POST(makeRequest());
@@ -267,13 +273,11 @@ describe("PR-API-11 이메일 재전송 Rate Limit 검증 (5회/1분)", () => {
     const response = await POST(makeRequest());
     const body = await response.json();
 
-    expect(response.status).not.toBe(429);
     expect(body.success).toBe(true);
-    expect(body.code).toBe(AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_SUCCESS);
     expect(resendVerificationEmail).toHaveBeenCalledTimes(1);
   });
 
-  // TC-04: 한도 초과 후 동일 윈도우 내 추가 요청 계속 차단
+  // TC-04: 한도 초과 후 계속 차단
   it("TC-04. 한도 초과 후 동일 윈도우 내 추가 요청은 계속 차단된다", async () => {
     for (let i = 0; i < 6; i++) {
       await POST(makeRequest());
@@ -286,27 +290,23 @@ describe("PR-API-11 이메일 재전송 Rate Limit 검증 (5회/1분)", () => {
 
       expect(response.status).toBe(429);
       expect(body.success).toBe(false);
-      expect(body.code).toBe(AUTH_API_CODES.RESEND_RATE_LIMIT_EXCEEDED);
     }
-
-    expect(resendVerificationEmail).toHaveBeenCalledTimes(0);
   });
 
-  // TC-05: 1분 경과 후 허용 (윈도우 리셋)
+  // TC-05: 1분 후 리셋
   it("TC-05. 1분 이상 경과 후 동일 이메일 요청이 다시 허용된다", async () => {
     for (let i = 0; i < 6; i++) {
       await POST(makeRequest());
     }
     vi.clearAllMocks();
 
+    // 시간 경과 시뮬레이션
     vi.advanceTimersByTime(61 * 1000);
 
     const response = await POST(makeRequest());
     const body = await response.json();
 
-    expect(response.status).not.toBe(429);
     expect(body.success).toBe(true);
-    expect(body.code).toBe(AUTH_API_CODES.EMAIL_VERIFICATION_RESEND_SUCCESS);
     expect(resendVerificationEmail).toHaveBeenCalledTimes(1);
   });
 });
